offline_sangeet/requirements.txt
asgiref==3.4.1
Django==3.2.16
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytube3==9.6.4
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.1.1
